Added Json dump function and troubleshooted Mac Function
I created this branch to test stuff out without messing with master/scripts folder. As I find errors that happen on my pc (let me know if you already fixed some of these I don't know for sure if I pulled the right version.

But added jsonFormatting(self, data) Takes data in and then uses json.dumps to format it.

get_mac_addresses() was causing issues on my machine, the output on one switch gave a empty line causing an index out of bound error but one time the error did not happen, so might need more exploring (i cant remember if the last number in a range(x,y) is inclusive or exclusive. so I added some lines to solve those problems to test out the Json function. Also if you are reading this while still on vacation I got the Pinnacle Audit logic to work and output to text file If we want to handle the output differently that's all good not set in stone on that way.